% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{addMakevarsToSpec}
\alias{addMakevarsToSpec}
\title{addMakevarsToSpec adds a line of CFLAGS, if needed}
\usage{
addMakevarsToSpec(specfile, makevar)
}
\arguments{
\item{specfile}{a specfile}

\item{makevar}{what to add to ~/.R/Makevars in specfile}
}
\value{
list of status "done" or "fail" and value, the specfile or NA
}
\description{
addMakevarsToSpec adds a line of CFLAGS, if needed
}
