% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{addPost}
\alias{addPost}
\title{addPost adds sections %post and %postun for
included libraries}
\usage{
addPost(specfile)
}
\arguments{
\item{specfile}{a specfile}
}
\value{
list of status "done" or "fail" and value, the specfile or NA
}
\description{
addPost adds sections %post and %postun for
included libraries
}
