% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{available.packages.OBS}
\alias{available.packages.OBS}
\title{This function generates an analogous matrix to available.packages()
but for all the packages in OBS}
\usage{
available.packages.OBS(obsproject = getOption("c2o.auto"), quiet = TRUE)
}
\arguments{
\item{obsproject}{Project in OBS where packages are taken from.}

\item{quiet}{If set to FALSE some progression info is given, default = TRUE.}
}
\value{
Matrix of all R packages avaialble in obsproject.
}
\description{
This function generates an analogous matrix to available.packages()
but for all the packages in OBS
}
