% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{buildforfiles}
\alias{buildforfiles}
\title{buildforfiles performs a build without %files section
and tries to contstruct all files from the build errors.}
\usage{
buildforfiles(
  pkg,
  pkgdir,
  specfile,
  localOBSdir = getOption("c2o.localOBSdir"),
  remoteproj = getOption("c2o.auto"),
  download.cache = getOption("c2o.download.cache"),
  ap = data.frame(available.packages(repos = getOption("c2o.cran")))
)
}
\arguments{
\item{pkg}{Name of a package in a R repo like CRAN}

\item{specfile}{Location of created specfile}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteproj}{Name of the OBS project}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\description{
buildforfiles performs a build without %files section
and tries to contstruct all files from the build errors.
}
