% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{buildinstallrpm}
\alias{buildinstallrpm}
\title{buildinstallrpm takes a R package name and tries to
build and locally install the corresponding rpm package}
\usage{
buildinstallrpm(packname, rpmbuildroot = "~/rpmbuild/")
}
\value{
No value
}
\description{
buildinstallrpm takes a R package name and tries to
build and locally install the corresponding rpm package
}
