% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check.settings}
\alias{check.settings}
\title{.onLoad prints the version number and checks settings of the package configuration
A warning is given, if settings are not complete!}
\usage{
check.settings()
}
\value{
No return value
check.settings validates all options settings

TRUE if build can start, FALSE otherwise, but tries to give hints.
}
\description{
.onLoad prints the version number and checks settings of the package configuration
A warning is given, if settings are not complete!
}
