% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanBuildlog}
\alias{cleanBuildlog}
\title{cleanBuildlog removes useless parts of the output
- timestamps get removed
- whitespace gets removed
- setup info from osc gets removed}
\usage{
cleanBuildlog(buildlog)
}
\arguments{
\item{buildlog}{array of lines resulting from call to osc}
}
\value{
cleaned up version of input
}
\description{
cleanBuildlog removes useless parts of the output
- timestamps get removed
- whitespace gets removed
- setup info from osc gets removed
}
