% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{cleanList}
\alias{cleanList}
\title{cleanList takes a package name and uses tools::package_dependencies
to construct either a list of OpenSUSE dependencies, which in CRAN
means the union of 'Depends', 'Imports' and 'LinkingTo', or
OpenSUSE recommends, which are represented in CRAN 'Suggests:'
Dependencies are constructed recursively, suggests not!
These lists are cleaned from installed base and Recommended packages
before returned.}
\usage{
cleanList(pkgname, kind)
}
\value{
Vector of CRAN package names
}
\description{
cleanList takes a package name and uses tools::package_dependencies
to construct either a list of OpenSUSE dependencies, which in CRAN
means the union of 'Depends', 'Imports' and 'LinkingTo', or
OpenSUSE recommends, which are represented in CRAN 'Suggests:'
Dependencies are constructed recursively, suggests not!
These lists are cleaned from installed base and Recommended packages
before returned.
}
