% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{cran2repo}
\alias{cran2repo}
\title{cran2repo tries to sync as much of a CRAN mirror
to an OBS repo as possible. Creating und Updating of packages
are done in one go in sequence along with increasing recursice
deplength. Usually the parameters are set using the options of
CRAN2OBS.}
\usage{
cran2repo(
  cran = getOption("c2o.cran"),
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  statusfile = getOption("c2o.statusfile"),
  download.cache = getOption("c2o.download.cache"),
  binary.cache = getOption("c2o.binary.cache"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{cran}{CRAN mirror to use}

\item{localOBS}{root directory to build remoteprj in locally}

\item{download.cache}{directory where source files from cran cached
sync status}

\item{binary.cache}{directory where newly built pacs are found}

\item{log}{logfile to use}

\item{repo}{OBS repo to use}
}
\value{
dataframe containing new syncstatus
}
\description{
cran2repo tries to sync as much of a CRAN mirror
to an OBS repo as possible. Creating und Updating of packages
are done in one go in sequence along with increasing recursice
deplength. Usually the parameters are set using the options of
CRAN2OBS.
}
