% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{CranOBSfromScratch}
\alias{CranOBSfromScratch}
\title{CranOBSfromScratch combines available.pacakges() and info of two OBS repos, i.e.
home:detlef:Automatic and d:l:R:released into a datafram combining all information
about the build status of the different packages.}
\usage{
CranOBSfromScratch(
  repo1 = "home:dsteuer:AutomaticCRAN",
  repo2 = "devel:languages:R:released"
)
}
\arguments{
\item{repo1, }{repo2 Repos containing automatic builds and manually improved builds resp.}
}
\value{
A dataframe like cleanDeps with additional infos for OBS packages
}
\description{
CranOBSfromScratch combines available.pacakges() and info of two OBS repos, i.e.
home:detlef:Automatic and d:l:R:released into a datafram combining all information
about the build status of the different packages.
}
