% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{CranOBSstatus}
\alias{CranOBSstatus}
\title{CranOBSstatus enriches cleanDeps for all packages in CRAN with their version numbers
in OBS}
\usage{
CranOBSstatus(quiet = TRUE, cran = NULL, obs = NULL)
}
\arguments{
\item{quiet}{If set to FALSE some progress info is given, default = TRUE.}

\item{cran}{If not NULL a matrix like returned from cleanDeps() must be given. If NULL cleanDeps()
is called.}

\item{obs}{If not NULL a matrix like returned from available.packages.OBS() must be given. If NULL
that function is called.}
}
\value{
A dataframe like cleanDeps with additional infos for OBS packages
}
\description{
CranOBSstatus enriches cleanDeps for all packages in CRAN with their version numbers
in OBS
}
