% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{createEmptySpec}
\alias{createEmptySpec}
\title{createEmptySpec takes the name of an R package and creates an specfile
with empty file section from its DESCRIPTION file.}
\usage{
createEmptySpec(
  packname,
  pacdir = file.path(getOption("c2o.localOBSdir"), getOption("c2o.auto"), paste0("R-",
    packname)),
  download.cache = getOption("c2o.download.cache"),
  cran = getOption("c2o.cran"),
  ap = data.frame(available.packages(repos = getOption("c2o.cran")))
)
}
\arguments{
\item{packname}{A package name of a package in a R repo like CRAN}

\item{download.cache}{Directory where sources are stored}

\item{ap}{A dataframe like the result of available.packages() or cleanDeps()
createEmptySpec uses Package, Version, and NeedsCompilation of that dataframe.}

\item{targetdir}{The directory where rpmbuild should do its work.
Sources of packages reside in rpmbuildroot/SOURCES, SPECS in rpmbuildroot/SPECS}
}
\value{
The path to the generated specfile, if creation succeeded,
a character string beginning with Failed: in case of an error.
}
\description{
createEmptySpec takes the name of an R package and creates an specfile
with empty file section from its DESCRIPTION file.
}
