% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{createNewOBSPackage}
\alias{createNewOBSPackage}
\title{This function creates an OBS package from an R package}
\usage{
createNewOBSPackage(obsproject, obscodir = "~/OBS/", pkg)
}
\arguments{
\item{obsproject}{Project in OBS where packages are taken from.}

\item{obscodir}{Directory where the local copy of the OBS package shall be created}
}
\description{
This function creates an OBS package from an R package
}
