% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{createOBSpac}
\alias{createOBSpac}
\title{createOBSpac takes the name of an R package and creates a 
package directory to be used in OBS.
The external tool osc is used to build
the OpenSUSE package. The resulting spec file will reside in a
package dir under a local checkout of a remote proj.
The %check section is deliberately empty. We rely
on the checks on CRAN.}
\usage{
createOBSpac(
  packname,
  localOBSdir = getOption("c2o.localOBSdir"),
  remoteproj = getOption("c2o.auto"),
  cran = getOption("c2o.cran"),
  ap = data.frame(available.packages(repos = getOption("c2o.cran")))
)
}
\arguments{
\item{packname}{Name of a package in a R repo like CRAN}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteproj}{Name of the OBS project}

\item{ap}{A dataframe containing info like from available.packages or cleanDeps.}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\value{
The path to the generated specfile.
}
\description{
createOBSpac takes the name of an R package and creates a 
package directory to be used in OBS.
The external tool osc is used to build
the OpenSUSE package. The resulting spec file will reside in a
package dir under a local checkout of a remote proj.
The %check section is deliberately empty. We rely
on the checks on CRAN.
}
