% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{depUnion}
\alias{depUnion}
\title{This function generates the complete set of dependencies, give a
vector of package names.
This is useful i.e. to build rstudio, to have a simple way
to generate the complete list of recursive dependencies.}
\usage{
depUnion(pkglist, ap = NULL)
}
\arguments{
\item{pkglist}{A vector of package names}

\item{ap}{A data structure like returned from cleanDeps, must
contain a column "recDep" containing recursive dependencies for
a given package. If NULL will be generated.}
}
\description{
This function generates the complete set of dependencies, give a
vector of package names.
This is useful i.e. to build rstudio, to have a simple way
to generate the complete list of recursive dependencies.
}
