% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{desc2spec}
\alias{desc2spec}
\title{desc2spec takes the name of an R package and generates a spec file
to be used in OBS. The external tool rpmbuild is used to build
the OpenSUSE package. The resulting spec file will reside where
rpmbuild puts it. The %check section is deliberately empty. We rely
on the checks on CRAN.}
\usage{
desc2spec(packname, rpmbuildroot = "~/rpmbuild/")
}
\arguments{
\item{packname}{A package name of a package in a R repo like CRAN}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\value{
The path to the generated specfile.
}
\description{
desc2spec takes the name of an R package and generates a spec file
to be used in OBS. The external tool rpmbuild is used to build
the OpenSUSE package. The resulting spec file will reside where
rpmbuild puts it. The %check section is deliberately empty. We rely
on the checks on CRAN.
}
