% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{.onLoad}
\alias{.onLoad}
\title{.onLoad prints the version number and checks settings of the package configuration
A warning is given, if settings are not complete!}
\usage{
.onLoad(libname, pkgname)
}
\value{
No return value
}
\description{
.onLoad prints the version number and checks settings of the package configuration
A warning is given, if settings are not complete!
}
