% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{dropFileSection}
\alias{dropFileSection}
\title{dropFileSection removes the filelist from a character array
representing a spec file. Keep the first line of the file list.}
\usage{
dropFileSection(speclines)
}
\arguments{
\item{speclines}{character vector with lines holding a specfile}
}
\value{
list of two components status and speclines
in status is "success" oder "fail", speclines either with empty file
section resp. unaltered.
}
\description{
dropFileSection removes the filelist from a character array
representing a spec file. Keep the first line of the file list.
}
