% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{extractDevelFilesFromLog}
\alias{extractDevelFilesFromLog}
\title{extractDevelFilesFromLog takes the lines of an osc build log and
extrats the list of files which need to be put in a devel packages.
Times must be stripped already.}
\usage{
extractDevelFilesFromLog(buildlog, pkg, systemRlib = "/usr/lib64/R/library/")
}
\arguments{
\item{buildlog}{Character vector containing the lines of an osc
build log}
}
\value{
Character vector containing entries for the files section of
a -devel package
}
\description{
extractDevelFilesFromLog takes the lines of an osc build log and
extrats the list of files which need to be put in a devel packages.
Times must be stripped already.
}
