% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{extractFilesFromSpec}
\alias{extractFilesFromSpec}
\title{extract the lines off of the %files section of a simple, i.e. non-devel
specfile}
\usage{
extractFilesFromSpec(specfile)
}
\arguments{
\item{specfile}{is a specfile}
}
\value{
vector of entries of %files section
}
\description{
extract the lines off of the %files section of a simple, i.e. non-devel
specfile
}
