% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{getOBSVersion2}
\alias{getOBSVersion2}
\title{getOBSVersion takes the name of an R package (from CRAN) and receives the corresponding
information off of OBS
This function is not exactly cheap run-time wise. Around 0.5s per package.}
\usage{
getOBSVersion2(pkg, obsproject = getOption("c2o.auto"))
}
\arguments{
\item{pkg}{name of an R package as found in CRAN}

\item{obsproject}{Project in OBS where packages are taken from.}
}
\value{
A list with components file containing the version number and
a boolean value hasDevel
}
\description{
getOBSVersion takes the name of an R package (from CRAN) and receives the corresponding
information off of OBS
This function is not exactly cheap run-time wise. Around 0.5s per package.
}
