% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logger}
\alias{logger}
\title{logger writes a character string in the console and append it to file
If msg is more than one line, only written to file}
\usage{
logger(msg, file = getOption("c2o.logfile"))
}
\arguments{
\item{msg}{character string}

\item{file}{to write to}
}
\value{
no return value
}
\description{
logger writes a character string in the console and append it to file
If msg is more than one line, only written to file
}
