% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{obsVersion}
\alias{obsVersion}
\title{obsVersion turns source version in OBS Version
(just for readability)}
\usage{
obsVersion(version)
}
\arguments{
\item{version}{character string representing CRAN version}
}
\value{
normalized OBS Version
}
\description{
obsVersion turns source version in OBS Version
(just for readability)
}
