% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pac.R
\name{pkg2pac}
\alias{pkg2pac}
\title{pkg2pac takes the name of an R package and creates or updates 
package directory to be used in OBS. 
The external tool osc is used to build
the OpenSUSE package. The resulting spec file will reside in a
package dir under a local checkout of a remote proj.}
\usage{
pkg2pac(
  pkg,
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  cran = getOption("c2o.cran"),
  statusfile = getOption("c2o.statusfile"),
  download.cache = getOption("c2o.download.cache"),
  binary.cache = getOption("c2o.binary.cache"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{Name of a package in a R repo like CRAN}

\item{localOBS}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteprj}{Name of the OBS project}

\item{cran}{CRAN mirror to use}

\item{statusfile}{a file holding current syncstate}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\value{
list of 'status', "done" or "fail", 'buildtype', 'value' set to NA
or character string and 'hasDevel'
}
\description{
pkg2pac takes the name of an R package and creates or updates 
package directory to be used in OBS. 
The external tool osc is used to build
the OpenSUSE package. The resulting spec file will reside in a
package dir under a local checkout of a remote proj.
}
