% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{pkg2repo}
\alias{pkg2repo}
\title{pk2repo does the same as cran2repo, but for a singl pkg}
\usage{
pkg2repo(
  pkg,
  cran = getOption("c2o.cran"),
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  statusfile = getOption("c2o.statusfile"),
  download.cache = getOption("c2o.download.cache"),
  binary.cache = getOption("c2o.binary.cache"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{cran}{CRAN mirror to use}

\item{localOBS}{root directory to build remoteprj in locally}

\item{download.cache}{directory where source files from cran cached
sync status}

\item{binary.cache}{directory where newly built pacs are found}

\item{log}{logfile to use}

\item{repo}{OBS repo to use}
}
\value{
dataframe containing new syncstatus
}
\description{
pk2repo does the same as cran2repo, but for a singl pkg
}
