% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{repoStatusCreate}
\alias{repoStatusCreate}
\title{repoStatusCreate creates a datafram from available.packages() and
available.packages.OBS() to have all the neccessary information to sync
the given CRAN mirror to the given OBS repo}
\usage{
repoStatusCreate(
  cran = getOption("c2o.cran"),
  repo = getOption("c2o.auto"),
  file = getOption("c2o.statusfile"),
  overwrite = FALSE
)
}
\arguments{
\item{cran}{is a CRAN mirror}

\item{repo}{is an OBS repo}

\item{file}{is the name of the file which will store the information}

\item{overwrite}{is a flag if an existing status file shall be overwritten}
}
\value{
a datafram containing the columns "Package", "Version", "License",
"NeedsCompilation", "recDep", "Suggests", "depLen", "OBSpkg", "File", "OBSVersion",
"triedVersion"
}
\description{
repoStatusCreate creates a datafram from available.packages() and
available.packages.OBS() to have all the neccessary information to sync
the given CRAN mirror to the given OBS repo
}
