% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{repoStatusfromScratch}
\alias{repoStatusfromScratch}
\title{repoStatusfromScratch creates a datafram from available.packages() and
available.packages.OBS() to have all the neccessary information to sync
the given CRAN mirror to the given OBS repo}
\usage{
repoStatusfromScratch(
  cran = getOption("c2o.cran"),
  repo = getOption("c2o.auto"),
  file = getOption("c2o.statusfile"),
  overwrite = FALSE
)
}
\arguments{
\item{cran}{is a CRAN mirror}

\item{repo}{is an OBS repo}

\item{file}{is the name of the file which will store the information}

\item{overwrite}{is a flag if an existing status file shall be overwritten}
}
\value{
a datafram containing the columns "Package", "Version", "License",
"NeedsCompilation", "recDep", "depLen", "OBSpkg", "File", "OBSVersion",
"triedVersion"
}
\description{
repoStatusfromScratch creates a datafram from available.packages() and
available.packages.OBS() to have all the neccessary information to sync
the given CRAN mirror to the given OBS repo
}
