% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pac.R
\name{resetpac}
\alias{resetpac}
\title{resetpac removes a pac completely and removes the
line from the status.
pkg will be handled as new in next status construction.}
\usage{
resetpac(
  pkg,
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  statusfile = getOption("c2o.statusfile"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{CRAN package name}

\item{localOBS}{base directory for osc checkouts}

\item{remoteprj}{the project for which packages are built}

\item{statusfile}{file holding the current sync status between
CRAN and remoteprj}

\item{cran}{CRAN mirror for download}
}
\value{
list "status" and "value"
}
\description{
resetpac removes a pac completely and removes the
line from the status.
pkg will be handled as new in next status construction.
}
