% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resetPkg}
\alias{resetPkg}
\title{resetPkg clears OBSVersion, triedVersion and hasDevel}
\usage{
resetPkg(pkg, status)
}
\arguments{
\item{pkg}{package name}

\item{status}{dataframe holding status}
}
\value{
updated dataframe holding status.
}
\description{
resetPkg clears OBSVersion, triedVersion and hasDevel
}
