% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{resetPkgRevDeps}
\alias{resetPkgRevDeps}
\title{resetPkgRevDeps clears OBSVersion, triedVersion and hasDevel 
from a package and all reverse dependencies of named package}
\usage{
resetPkgRevDeps(pkg, status)
}
\arguments{
\item{pkg}{package name}

\item{status}{dataframe holding status}
}
\value{
updated dataframe holding status.
}
\description{
resetPkgRevDeps clears OBSVersion, triedVersion and hasDevel 
from a package and all reverse dependencies of named package
}
