% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{resetStatusOfpkg}
\alias{resetStatusOfpkg}
\title{resetStatusOfpkg removes entries for OBSVersion, hasDevel and
triedVersion}
\usage{
resetStatusOfpkg(
  pkg,
  status = NA,
  statusfile = getOption("c2o.statusfile"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{a packagename}

\item{status}{a dataframe containing current status. If NA statusfile
will be read}

\item{statusfile}{file containing status information about ongoing sync}

\item{log}{logfile}
}
\value{
dataframe holding updated status
}
\description{
resetStatusOfpkg removes entries for OBSVersion, hasDevel and
triedVersion
}
