% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{rmExeFromDoc}
\alias{rmExeFromDoc}
\title{rmExeFromDoc removes wrongly set exe bits from
documentation files}
\usage{
rmExeFromDoc(specfile, exefiles)
}
\arguments{
\item{specfile}{a specfile}

\item{exefiles}{vector of fq-filenames}
}
\value{
list of status "done" or "fail" and value, the specfile or NA
}
\description{
rmExeFromDoc removes wrongly set exe bits from
documentation files
}
