% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{setupOBSdir}
\alias{setupOBSdir}
\title{setupOBSdir creates the OBS project for pkg on the local machine
or checks out the copy from remoteprj.}
\usage{
setupOBSdir(
  pkg,
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  cran = getOption("c2o.cran"),
  status = getOption("c2o.status"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{CRAN package name}

\item{localOBS}{base directory for osc checkouts}

\item{remoteprj}{the project for which packages are built}

\item{cran}{CRAN mirror for download}

\item{status}{dataframe holding the current sync status between
CRAN and remoteprj}
}
\value{
list( status, problem) where 'status' is "done" or "fail"
and 'problem' is NA or a string with an error message
}
\description{
setupOBSdir creates the OBS project for pkg on the local machine
or checks out the copy from remoteprj.
}
