% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pac.R
\name{setuppac}
\alias{setuppac}
\title{setuppac creates the OBS project for pkg on the local machine
or checks out the copy from remoteprj.}
\usage{
setuppac(
  pkg,
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  cran = getOption("c2o.cran"),
  status = getOption("c2o.status"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{CRAN package name}

\item{localOBS}{base directory for osc checkouts}

\item{remoteprj}{the project for which packages are built}

\item{cran}{CRAN mirror for download}

\item{status}{dataframe holding the current sync status between
CRAN and remoteprj}
}
\value{
list( status, value) where 'status' is "done" or "fail"
and 'value' is the created directory or a string with an error message
}
\description{
setuppac creates the OBS project for pkg on the local machine
or checks out the copy from remoteprj.
}
