% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{showPkg}
\alias{showPkg}
\title{showPkg return the line of status file for pkg}
\usage{
showPkg(pkg, status = readStatus(getOption("c2o.statusfile")))
}
\arguments{
\item{pkg}{pkgname}

\item{status}{datafram conataining sync state}
}
\description{
showPkg return the line of status file for pkg
}
