% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{sysreq2depends}
\alias{sysreq2depends}
\title{sysreq2depends generats Requires and BuildRequires
from SystemRequirements: lines in DESCRIPTION}
\usage{
sysreq2depends(line)
}
\arguments{
\item{line}{the line from DESCRIPTION}
}
\value{
list with components 'sysdepends' and 'sysbuilddepends'
}
\description{
sysreq2depends generats Requires and BuildRequires
from SystemRequirements: lines in DESCRIPTION
}
