% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{testbuild}
\alias{testbuild}
\title{testbuild performs a testbuild and analyses the result}
\usage{
testbuild(
  pkg,
  pac,
  specfile,
  log = getOption("c2o.logfile"),
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  download.cache = getOption("c2o.download.cache"),
  binary.cache = getOption("c2o.binary.cache"),
  ap = if (!is.null(getOption("c2o.status"))) getOption("c2o.status")
)
}
\arguments{
\item{pkg}{Name of a package in a R repo like CRAN}

\item{specfile}{Location of created specfile}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteproj}{Name of the OBS project}
}
\value{
list of status, value, buildlog
}
\description{
testbuild performs a testbuild and analyses the result
}
