% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{updateOBSPackage}
\alias{updateOBSPackage}
\title{This function updates a project already existing in OBS}
\usage{
updateOBSPackage(obsproject, obscodir = "~/OBS/", pkg)
}
\arguments{
\item{obsproject}{Project in OBS where packages are taken from.}

\item{obscodir}{Directory where the local copy of the OBS package shall be created}

\item{pkg}{Package name}
}
\description{
This function updates a project already existing in OBS
}
