% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{updateStatusOfpkg}
\alias{updateStatusOfpkg}
\title{updateStatusOfpkg incorporates new information about a tried build in
the status dataframe.}
\usage{
updateStatusOfpkg(
  status,
  pkg,
  syncresult,
  always.save = TRUE,
  statusfile = getOption("c2o.statusfile"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{status}{datframe holding status information about remoteprj}

\item{pkg}{CRAN package name which has new information}

\item{syncresult}{holds a list with components status, version and hasDevel}

\item{statusfile}{file containing status information about ongoing sync}

\item{log}{logfile}

\item{always.safe}{boolean that defines if statusfile should be written for
every package}
}
\value{
updated status dataframe
}
\description{
updateStatusOfpkg incorporates new information about a tried build in
the status dataframe.
}
