% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pac.R
\name{uploadpac}
\alias{uploadpac}
\title{uploadpac pushes a successfully built pac to remoteprj}
\usage{
uploadpac(
  pkg,
  version,
  buildtype,
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{CRAN package name}

\item{version}{to upload}

\item{buildtype}{"initial build" or "update"}

\item{localOBS}{base directory for osc checkouts}

\item{remoteprj}{the project for which packages are built}

\item{cran}{CRAN mirror for download}

\item{status}{dataframe holding the current sync status between
CRAN and remoteprj}
}
\value{
list "status" and "value"
}
\description{
uploadpac pushes a successfully built pac to remoteprj
}
