% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{writeStatus}
\alias{writeStatus}
\title{writeStatus writes status dataframe to statusfile}
\usage{
writeStatus(status, statusfile = getOption("c2o.statusfile"))
}
\arguments{
\item{status}{dataframe with status info}

\item{statusfile}{file to save info into}
}
\description{
writeStatus writes status dataframe to statusfile
}
