% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{buildforfiles}
\alias{buildforfiles}
\title{buildforfiles performs a build without %files section
and tries to contstruct all files from the build errors.}
\usage{
buildforfiles(
  pkg,
  pac,
  specfile,
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  download.cache = getOption("c2o.download.cache"),
  binary.cache = getOption("c2o.binary.cache"),
  ap = getOption("c2o.status"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{Name of a package in a R repo like CRAN}

\item{pac}{directory for OBS}

\item{specfile}{Location of created specfile}

\item{remoteprj}{Name of the OBS project}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}
}
\description{
buildforfiles performs a build without %files section
and tries to contstruct all files from the build errors.
}
