% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{buildinstallrpm}
\alias{buildinstallrpm}
\title{buildinstallrpm takes a R package name and tries to
build and locally install the corresponding rpm package}
\usage{
buildinstallrpm(packname, rpmbuildroot = "~/rpmbuild/")
}
\arguments{
\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\value{
No value
}
\description{
#' @param packname A package name of a package in a R repo like CRAN
}
