% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{check.settings}
\alias{check.settings}
\title{check.settings validates all options settings}
\usage{
check.settings()
}
\value{
TRUE if build can start, FALSE otherwise, but tries to give hints.
}
\description{
check.settings validates all options settings
}
