% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{cleanDeps}
\alias{cleanDeps}
\title{cleanDeps returns a structure like available.package()
but including recursive dependencies (Depends, Imports, LinkingTo).
Because recommended packages may be assumed to be installed
in a OpenSUSE installation with R, these are removed together with
base packages from the dependencies.}
\usage{
cleanDeps(repo = getOption("c2o.cran"))
}
\arguments{
\item{repo}{to build the dependency matrix off}
}
\value{
A dataframe containing from available.packages() columns "Package",
"Version", "License", "NeedsCompilation", a cleaned-up version of "Suggests",
and new columns 'recDep', containing all recursive dependencies, and 'depLen'
containing the number of recursive dependencies.
}
\description{
cleanDeps returns a structure like available.package()
but including recursive dependencies (Depends, Imports, LinkingTo).
Because recommended packages may be assumed to be installed
in a OpenSUSE installation with R, these are removed together with
base packages from the dependencies.
}
