% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{cleanList}
\alias{cleanList}
\title{cleanList takes a package name and uses tools::package_dependencies
to construct either a list of OpenSUSE dependencies, which in CRAN
means the union of 'Depends', 'Imports' and 'LinkingTo', or
OpenSUSE recommends, which are represented in CRAN 'Suggests:'
Dependencies are constructed recursively, suggests not!
These lists are cleaned from installed base and recommended packages
before returned.}
\usage{
cleanList(pkgname, kind, repo = getOption("c2o.cran"))
}
\arguments{
\item{pkgname}{name of CRAN package}

\item{kind}{of list: "depends" or "suggests"}

\item{repo}{repo to build list from}
}
\value{
Vector of CRAN package names
}
\description{
cleanList takes a package name and uses tools::package_dependencies
to construct either a list of OpenSUSE dependencies, which in CRAN
means the union of 'Depends', 'Imports' and 'LinkingTo', or
OpenSUSE recommends, which are represented in CRAN 'Suggests:'
Dependencies are constructed recursively, suggests not!
These lists are cleaned from installed base and recommended packages
before returned.
}
