% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{cran2repo}
\alias{cran2repo}
\title{cran2repo tries to sync as much of a CRAN mirror
to an OBS repo as possible. Creating und Updating of packages
are done in one go in sequence along with increasing recursice
deplength. Usually the parameters are set using the options of
CRAN2OBS.}
\usage{
cran2repo(
  cran = getOption("c2o.cran"),
  localOBS = getOption("c2o.localOBSdir"),
  remoteprj = getOption("c2o.auto"),
  status = getOption("c2o.status"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{cran}{CRAN mirror to use}

\item{localOBS}{root directory to build remoteprj in locally}

\item{status}{dataframe containing status information about}

\item{log}{logfile to use}

\item{repo}{OBS repo to use}

\item{binary.cache}{directory where newly built pacs are found}

\item{download.cache}{directory where source files from cran cached
sync status}
}
\description{
cran2repo tries to sync as much of a CRAN mirror
to an OBS repo as possible. Creating und Updating of packages
are done in one go in sequence along with increasing recursice
deplength. Usually the parameters are set using the options of
CRAN2OBS.
}
