% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{CranOBSfromScratch}
\alias{CranOBSfromScratch}
\title{CranOBSfromScratch combines available.packages() and info of two OBS repos, i.e.
home:detlef:AutomaticCRAN and d:l:R:released into a datafram combining all information
about the build status of the different packages.}
\usage{
CranOBSfromScratch(
  cran = getOption("c2o.cran"),
  repo1 = getOption("c2o.auto"),
  repo2 = getOption("c2o.manual")
)
}
\arguments{
\item{repo1}{and}

\item{repo2}{Repos containing automatic builds and manually improved builds resp.}
}
\value{
A dataframe like cleanDeps with additional infos for OBS packages
}
\description{
CranOBSfromScratch combines available.packages() and info of two OBS repos, i.e.
home:detlef:AutomaticCRAN and d:l:R:released into a datafram combining all information
about the build status of the different packages.
}
