% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{createEmptySpec}
\alias{createEmptySpec}
\title{createEmptySpec takes the name of an R package and creates an specfile
with empty file section from its DESCRIPTION file.}
\usage{
createEmptySpec(
  packname,
  rpmbuildroot = "~/rpmbuild/",
  ap = data.frame(available.packages(repos = "https://cloud.r-project.org"))
)
}
\arguments{
\item{packname}{A package name of a package in a R repo like CRAN}

\item{rpmbuildroot}{The directory where rpmbuild should do its work.
Sources of packages reside in rpmbuildroot/SOURCES, SPECS in rpmbuildroot/SPECS}

\item{ap}{A dataframe like the result of available.packages() or cleanDeps()
createEmptySpec uses Version and NeedsCompilation of that dataframe.}
}
\value{
The path to the generated specfile, if creation succeeded,
a character string beginning with Failed: in case of an error.
}
\description{
createEmptySpec takes the name of an R package and creates an specfile
with empty file section from its DESCRIPTION file.
}
