% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{createEmptySpec}
\alias{createEmptySpec}
\title{createEmptySpec takes the name of an R package and creates an specfile
with empty file section from its DESCRIPTION file. The OBS directory must
be prepared by setupOBSdir}
\usage{
createEmptySpec(
  pkg,
  pac = file.path(getOption("c2o.localOBSdir"), getOption("c2o.auto"), paste0("R-",
    pkg)),
  download.cache = getOption("c2o.download.cache"),
  cran = getOption("c2o.cran"),
  status = getOption("c2o.status"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{pkg}{A package name of a package in a R repo like CRAN}

\item{download.cache}{Directory where sources are stored}

\item{status}{A dataframe like the result of available.packages() or cleanDeps()
createEmptySpec uses Package, Version, and NeedsCompilation of that dataframe.}

\item{targetdir}{The directory where rpmbuild should do its work.
Sources of packages reside in rpmbuildroot/SOURCES, SPECS in rpmbuildroot/SPECS}
}
\value{
list of status, 'fail' and 'done' and value. 'value' is either the filename of the generated specfile
or a problem description
}
\description{
createEmptySpec takes the name of an R package and creates an specfile
with empty file section from its DESCRIPTION file. The OBS directory must
be prepared by setupOBSdir
}
