% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{createOBSpac}
\alias{createOBSpac}
\title{createOBSspec takes the name of an R package and creates a 
spec file to be used in OBS.
The external tool osc is used to build
the OpenSUSE package. The resulting spec file will reside in a
package dir under a local checkout of a remote proj.
The %check section is deliberately empty. We rely
on the checks on CRAN.}
\usage{
createOBSpac(
  packname,
  localOBSdir = "~/OBS",
  remoteproj = "home:dsteuer:AutomaticCRAN",
  ap = data.frame(available.packages(repos = "https://cloud.r-project.org"))
)
}
\arguments{
\item{packname}{A package name of a package in a R repo like CRAN}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteproj}{Name of the OBS project}

\item{ap}{A dataframe containing info like from available.packages or cleanDeps.}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\value{
The path to the generated specfile.
}
\description{
createOBSspec takes the name of an R package and creates a 
spec file to be used in OBS.
The external tool osc is used to build
the OpenSUSE package. The resulting spec file will reside in a
package dir under a local checkout of a remote proj.
The %check section is deliberately empty. We rely
on the checks on CRAN.
}
