% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specs.R
\name{expandSpecForDevel}
\alias{expandSpecForDevel}
\title{expandSpecForDevel takes a specfile and adds a -devel section}
\usage{
expandSpecForDevel(specfile, mainfiles, develfiles)
}
\arguments{
\item{specfile}{A spec file}

\item{mainfiles}{character vector with entries for the files section}

\item{develfiles}{character vector with entries for the files -devel section}
}
\value{
list of status "done" or "fail" and value, the specfile or NA
}
\description{
expandSpecForDevel takes a specfile and adds a -devel section
}
