% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{getOBSVersion}
\alias{getOBSVersion}
\title{This function takes the name of an R package (from CRAN) and receives the corresponding
information off OBS
This function is not exactly cheap run-time wise. Around 0.5s per package.}
\usage{
getOBSVersion(pkg, quiet = TRUE)
}
\arguments{
\item{pkg}{A character string containing the name of a R package as found in CRAN}

\item{quiet}{If set to FALSE some progression info is given, default = TRUE.}
}
\value{
A list with components file containing the source file name and
version containing the version.
}
\description{
This function takes the name of an R package (from CRAN) and receives the corresponding
information off OBS
This function is not exactly cheap run-time wise. Around 0.5s per package.
}
