% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{readStatus}
\alias{readStatus}
\title{readStatus reads a statusfile and return the datafram}
\usage{
readStatus(statusfile = getOption("c2o.statusfile"))
}
\arguments{
\item{statusfile}{filename of statusfile}
}
\value{
status datafram
}
\description{
readStatus reads a statusfile and return the datafram
}
