% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{showOBSstatus}
\alias{showOBSstatus}
\title{This functions shows a table for all packages in OBS with version numbers
of CRAN and OBS
return OBSstatus A dataframe containing esp. version information from CRAN for all OBS packages}
\usage{
showOBSstatus()
}
\description{
This functions shows a table for all packages in OBS with version numbers
of CRAN and OBS
return OBSstatus A dataframe containing esp. version information from CRAN for all OBS packages
}
