% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{statusOBS}
\alias{statusOBS}
\title{statusOBS checks if a package already exists either as remote pac on build.opensuse.org}
\usage{
statusOBS(pkg, remoteproj = getOption("c2o.auto"))
}
\arguments{
\item{pkg}{CRAN R package name}

\item{remoteproj}{Existing project to build packages for in OBS (build.opensuse.org)}
}
\value{
Version number of package in OBS or NA
}
\description{
statusOBS checks if a package already exists either as remote pac on build.opensuse.org
}
