% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{statusOBS}
\alias{statusOBS}
\title{statusOBS checks if a package already exists either as local
OBS pac or remote pac on build.opensuse.org}
\usage{
statusOBS(pkg, remoteproj = "home:dsteuer:AutomaticCRAN")
}
\arguments{
\item{remoteproj}{Existing project to buid packages for in OBS (build.opensuse.org)}

\item{one}{packname CRAN R package name}
}
\value{
Version number of package in OBS or NA
}
\description{
statusOBS checks if a package already exists either as local
OBS pac or remote pac on build.opensuse.org
}
