% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{testbuild}
\alias{testbuild}
\title{testbuild performs a testbuild and analyses the result}
\usage{
testbuild(
  pkg,
  pkgdir,
  specfile,
  logfile = "buildlog",
  localOBSdir = "~/OBS",
  remoteproj = "home:dsteuer:AutomaticCRAN",
  download.cache = "~/rpmbuild/SOURCES"
)
}
\arguments{
\item{pkg}{Name of a package in a R repo like CRAN}

\item{specfile}{Location of created specfile}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteproj}{Name of the OBS project}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\description{
testbuild performs a testbuild and analyses the result
}
