% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{updateOBSpkg}
\alias{updateOBSpkg}
\title{updateOBSpkg takes the name of an R package with newer
version CRAN and tries to update with minimal changes to
the existing spec.}
\usage{
updateOBSpkg(
  pkg,
  localOBSdir = getOption("c2o.localOBSdir"),
  remoteproj = getOption("c2o.auto"),
  download.cache = getOption("c2o.download.cache"),
  cran = getOption("c2o.cran"),
  ap = data.frame(available.packages(repos = getOption("c2o.cran")))
)
}
\arguments{
\item{pkg}{Name of a package in a R repo like CRAN}

\item{localOBSdir}{The top level directory of a checkout of the project
you want to create your package in.}

\item{remoteproj}{Name of the OBS project}

\item{ap}{A dataframe containing info like from available.packages or cleanDeps.}

\item{rpmbuildroot}{The directory where rpmbuild should do its work}
}
\value{
The path to the generated specfile.
}
\description{
updateOBSpkg takes the name of an R package with newer
version CRAN and tries to update with minimal changes to
the existing spec.
}
