% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{updateStatusOfpkg}
\alias{updateStatusOfpkg}
\title{updateStatusOfpkg incorporates new information about a tried build in
the status dataframe.}
\usage{
updateStatusOfpkg(
  status,
  pkg,
  syncresult,
  always.save = TRUE,
  file = getOption("c2o.statusfile"),
  log = getOption("c2o.logfile")
)
}
\arguments{
\item{status}{the dataframe containing status information about remoteprj}

\item{pkg}{CRAN package name which has new information}

\item{version}{latest version tried}

\item{success}{if or not the build was successful}
}
\value{
updated status dataframe
}
\description{
updateStatusOfpkg incorporates new information about a tried build in
the status dataframe.
}
