% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc2spec.R
\name{versionFromTgz}
\alias{versionFromTgz}
\title{versionFromTgz (deprecated) extracts the version info, normalizes it and return
a charater string containing that version}
\usage{
versionFromTgz(pkg, tgzname)
}
\arguments{
\item{pkg}{CRAN package name}

\item{tgzname}{filename to extract version from}
}
\value{
Version string
}
\description{
versionFromTgz (deprecated) extracts the version info, normalizes it and return
a charater string containing that version
}
